+++ lib/cleanup.sh
@@ -1,5 +1,6 @@
 #!/usr/bin/env bash
 source "$(dirname "${BASH_SOURCE[0]}")/cleanup_empty_dirs.sh"
+source "$(dirname "${BASH_SOURCE[0]}")/cleanup_junk_files.sh"
 
 ptk_cleanup_library() {
     local path="$1"
@@ -13,4 +14,8 @@ ptk_cleanup_library() {
     echo
     echo "Searching empty directories..."
     ptk_find_empty_dirs "$path"
+
+    echo
+    echo "Searching junk files..."
+    ptk_find_junk_files "$path"
 }